Updating inline JavaDoc.

+++ src/test/java/scheduler/InMemoryRepository.java
@@ -3,22 +3,22 @@ package scheduler;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class InMemoryRepository implements StreamingRepository<Cart<Double>, Double> {
+public class InMemoryRepository implements StreamingRepository<Cart<Integer>, Integer> {
 
-    private final Map<Long, Cart<Double>> repo =  new ConcurrentHashMap<>();
+    private final Map<Long, Cart<Integer>> repo =  new ConcurrentHashMap<>();
 
     @Override
-    public Cart<Double> getById(Long id) {
+    public Cart<Integer> getById(Long id) {
         return repo.get(id);
     }
 
     @Override
-    public void save(Cart<Double> item) {
+    public void save(Cart<Integer> item) {
         repo.put(item.getId(), item);
     }
 
     @Override
-    public Order<Double> saveOrder(Order<Double> order) {
+    public Order<Integer> saveOrder(Order<Integer> order) {
         return order;
     }
